chore: memberApply 파일 분류 #86
