+++ src/main/java/com/example/olebackend/service/MemberService.java
@@ -27,8 +27,6 @@ public class MemberService {
 
     private final MemberRepository memberRepository;
 //    private final PasswordEncoder passwordEncoder;
-    private final LessonRepository lessonRepository;
-    private final MemberApplyRepository memberApplyRepository;
 
     public void signUp(MemberSignUpRequest memberSignUpDto) throws Exception{
 
@@ -55,39 +53,5 @@ public class MemberService {
 
 //        member.passwordEncode(passwordEncoder);
         memberRepository.save(member);
-    }
-
-
-
-    public MemberApply applyLesson(Long lessonId,Long memberId) {
-
-        Lesson lesson=lessonRepository.findById(lessonId)
-                .orElseThrow(() ->new GeneralException(LESSON_NOT_FOUND));
-
-        Member member=memberRepository.findById(memberId)
-                .orElseThrow(() ->new GeneralException(MEMBER_NOT_FOUND));
-
-        if(memberApplyRepository.existsByLessonIdAndMemberId(lessonId, memberId)){
-            throw new GeneralException(LESSONAPPLY_ALREADY_EXISTS);
-        }
-
-        MemberApply memberApply = MemberConverter.toMemberApply(lesson,member);
-
-        return memberApplyRepository.save(memberApply);
-    }
-
-    @Transactional
-    public void cancelLesson(Long lessonId,Long memberId) {
-
-        Optional<Lesson> optionalLesson = lessonRepository.findById(lessonId);
-        if (optionalLesson.isEmpty()) {
-            throw new GeneralException(LESSON_NOT_FOUND);
-        }
-
-        if (!memberApplyRepository.existsByLessonIdAndMemberId(lessonId, memberId)) {
-            throw new GeneralException(LESSONAPPLY_NOT_FOUND);
-        }
-
-        memberApplyRepository.removeByLessonIdAndMemberId(lessonId, memberId);
     }
 }
