+++ src/main/java/com/example/olebackend/web/dto/MemberApplyResponse.java
@@ -1,0 +1,16 @@
+package com.example.olebackend.web.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+public class MemberApplyResponse {
+    @Builder
+    @Getter
+    @NoArgsConstructor
+    @AllArgsConstructor
+    public static class applyResultDTO{
+        Long memberApplyId;
+    }
+}
