+++ src/main/java/com/example/olebackend/web/dto/MemberResponse.java
@@ -30,12 +30,4 @@ public class MemberResponse {
     public static class JoinResultDTO{
         Long memberId;
     }
-
-    @Builder
-    @Getter
-    @NoArgsConstructor
-    @AllArgsConstructor
-    public static class applyResultDTO{
-        Long memberApplyId;
-    }
 }
