+++ src/main/resources/application.yml
@@ -20,7 +20,7 @@ spring:
         format_sql: true
         use_sql_comments: true
         hbm2ddl:
-          auto: create
+          auto: none
         default_batch_fetch_size: 1000${}
 
     defer-datasource-initialization: true  # 초기 데이터 적재 설정
